requirements.txt
mysql-connector
pycryptodome
tqdm
socket
os
flask